Add explanatory text to privacy screen (#690)
This also includes a refactor on the ServerRules-screen. It's built the same way
