+++ Mastodon/Scene/Onboarding/ServerRules/ServerRuleSection.swift
@@ -10,7 +10,6 @@ import MastodonAsset
 import MastodonLocalization
 
 enum ServerRuleSection: Hashable {
-    case header
     case rules
 }
 
@@ -20,10 +19,6 @@ extension ServerRuleSection {
     ) -> UITableViewDiffableDataSource<ServerRuleSection, ServerRuleItem> {
         return UITableViewDiffableDataSource(tableView: tableView) { tableView, indexPath, item in
             switch item {
-            case .header(let domain):
-                let cell = tableView.dequeueReusableCell(withIdentifier: String(describing: OnboardingHeadlineTableViewCell.self), for: indexPath) as! OnboardingHeadlineTableViewCell
-                cell.subTitleLabel.text = L10n.Scene.ServerRules.subtitle(domain)
-                return cell
             case .rule(let ruleContext):
                 let cell = tableView.dequeueReusableCell(withIdentifier: String(describing: ServerRulesTableViewCell.self), for: indexPath) as! ServerRulesTableViewCell
                 cell.indexImageView.image = UIImage(systemName: "\(ruleContext.index + 1).circle") ?? UIImage(systemName: "questionmark.circle")
